adição do put e delete ao questoescontroller, adição de validação de token em delete, TBA: Em todos os endpoints

+++ src/main/java/com/alphasolutions/eventapi/service/QuestoesService.java
@@ -1,8 +1,8 @@
 package com.alphasolutions.eventapi.service;
 
 import com.alphasolutions.eventapi.model.Questoes;
+import com.alphasolutions.eventapi.model.QuestoesDTO;
 import com.alphasolutions.eventapi.repository.QuestoesRepository;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -11,8 +11,11 @@ import java.util.Optional;
 @Service
 public class QuestoesService {
 
-    @Autowired
-    private QuestoesRepository questoesRepository;
+    private final QuestoesRepository questoesRepository;
+
+    public QuestoesService(QuestoesRepository questoesRepository) {
+        this.questoesRepository = questoesRepository;
+    }
 
     public List<Questoes> findAll() {
         return questoesRepository.findAll();
@@ -26,8 +29,24 @@ public class QuestoesService {
         return questoesRepository.save(questoes);
     }
 
-    public void deleteById(Long id) {
-        questoesRepository.deleteById(id);
+    public void deleteById(Long id) throws Exception {
+        if (questoesRepository.existsById(id)) {
+            questoesRepository.deleteById(id);
+        }else{
+            throw new Exception("Não existe esta questão!");
+        }
+
+    }
+
+    public void updateQuestoes(QuestoesDTO questoes) throws Exception {
+        Questoes questao = questoesRepository.findById(questoes.getIdQuestao()).orElse(null);
+        if(questao == null) {
+            throw new Exception("Questão não encontrada");
+        }
+
+        Questoes newQuestion = new Questoes(questoes.getIdQuestao(),questoes.getEnunciado(),questoes.getChoices(),questoes.getCorrectAnswer(),questoes.getIdPalestra());
+        questoesRepository.save(newQuestion);
+
     }
 
     public List<Questoes> findQuestoesByPalestra(Long idPalestra) {
